requirements.txt
asgiref==3.8.1
Django==4.2.20
django-extensions==3.2.3
djangorestframework==3.15.2
Faker==37.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
psycopg2-binary
